gp2gp-translator/src/test/java/uk/nhs/adaptors/pss/translator/service/BundleMapperServiceTest.java
package uk.nhs.adaptors.pss.translator.service;

import static org.assertj.core.api.Assertions.assertThat;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.when;
import static org.springframework.util.ResourceUtils.getFile;

import static uk.nhs.adaptors.pss.translator.util.XmlUnmarshallUtil.unmarshallFile;

import java.util.ArrayList;
import java.util.List;

import org.checkerframework.common.value.qual.MinLen;
import org.hl7.fhir.dstu3.model.Bundle;
import org.hl7.fhir.dstu3.model.DomainResource;
import org.hl7.fhir.dstu3.model.Organization;
import org.hl7.fhir.dstu3.model.Patient;
import org.hl7.fhir.dstu3.model.Reference;
import org.hl7.v3.RCMRIN030000UK06Message;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.junit.runner.RunWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.Mockito;
import org.mockito.Spy;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;

import lombok.SneakyThrows;
import uk.nhs.adaptors.pss.translator.generator.BundleGenerator;
import uk.nhs.adaptors.pss.translator.mapper.AgentDirectoryMapper;
import uk.nhs.adaptors.pss.translator.mapper.LocationMapper;
import uk.nhs.adaptors.pss.translator.mapper.PatientMapper;

@RunWith(SpringJUnit4ClassRunner.class)
public class BundleMapperServiceTest {

    private static final String XML_RESOURCES_BASE = "xml/RCMRIN030000UK06/";

    @Autowired
    private BundleMapperService bundleMapperService;

    @Test
    public void testBundleIsCorrectlyMapped() {
        RCMRIN030000UK06Message xml = unmarshallCodeElement("structuredRecord.xml");
        Bundle bundle = bundleMapperService.mapToBundle(xml);
        assertThat(bundle.getEntry()).isEmpty();
        //@TODO: Think of what to test later
    }

    @SneakyThrows
    private RCMRIN030000UK06Message unmarshallCodeElement(String fileName) {
        return unmarshallFile(getFile("classpath:" + XML_RESOURCES_BASE + fileName), RCMRIN030000UK06Message.class);
    }
}
